Baseflow Channel Removal and Documentation

+++ data-raw/scripts/hydraulic-training-data-vector.R
@@ -1,6 +1,8 @@
 library(tidyverse)
 library(sf)
-
+library(future)
+library(future.apply)
+library(furrr)
 #source(here::here("data-raw", "scripts", "data-functions.R"))
 #source(here::here("data-raw", "scripts", "suitability-functions.R"))
 library(habistat)
@@ -50,11 +52,12 @@ stan_csvs <- tribble(~flow_cfs, ~filename,
                       5000, "5000cfs_071113.csv.gz") |>
   mutate(filename = file.path(stan_dir, filename))
 
-
 stan_hsi_by_flow <-
   stan_csvs |>
   vector_import_srh2d(units = "m") |>
   vector_calculate_hsi(hsi_func = vector_dvhsi_hqt)
+
+### ORIGINAL VERSION
 
 stan_mesh_prepped <-
   stan_thiessen |>
@@ -69,6 +72,45 @@ stan_result <-
 outpath <- here::here("data-raw", "results", "fsa_stan.Rds")
 
 stan_result |> saveRDS(outpath)
+
+### VARIATION FOR REARING - BASEFLOW REMOVAL
+
+# define baseflow mask as 500 cfs (COMID 2819818 observed dry season baseflow is 306 cfs)
+stan_baseflow_poly <- stan_thiessen |>
+  filter(vid %in% filter(stan_hsi_by_flow, inundated & flow_cfs == 500)$vid) |>
+  summarize() |>
+  st_union()
+
+stan_mesh_prepped_nbfc <-
+  stan_thiessen |>
+  vector_prep_mesh(group_polygons = stan_comid,
+                   mask_polygons = stan_baseflow_poly,
+                   .group_var = comid,
+                   .id_var = vid)
+
+stan_result_nbfc <-
+  vector_summarize_hsi(stan_mesh_prepped_nbfc, stan_hsi_by_flow) |>
+  suitability_postprocess(stan_comid, .group_var = comid)
+
+outpath <- here::here("data-raw", "results", "fsa_stan_nbfc.Rds")
+
+stan_result_nbfc |> saveRDS(outpath)
+
+### VARIATION FOR SPAWNING - ALTERNATIVE HSI
+
+stan_hsi_by_flow_spawning <-
+  stan_csvs |>
+  vector_import_srh2d(units = "m") |>
+  vector_calculate_hsi(hsi_func = vector_dvhsi_spawning)
+
+stan_result_spawning <-
+  vector_summarize_hsi(stan_mesh_prepped, stan_hsi_by_flow_spawning) |>
+  suitability_postprocess(stan_comid, .group_var = comid)
+
+outpath <- here::here("data-raw", "results", "fsa_stan_spawning.Rds")
+
+stan_result_spawning |> saveRDS(outpath)
+
 
 # YUBA RIVER -------------------------------------------------------------
 
@@ -109,13 +151,16 @@ yuba_hsi_by_flow_by_reach <-
 
 yuba_hsi_by_flow_by_reach |> saveRDS(file.path(yuba_dir, "yuba_hsi_by_flow_by_reach.Rds"))
 
-yuba_mesh_prepped_by_reach <-
+yuba_datasets <-
   tribble(~reach, ~vertices, ~thiessen,
-    "EDR", "1t_oT37auM_25SnU3TLrXa_xCyoV9H9qt", "1s-FhTTDlclwcuu-H-9wIwwC1K7lEijXI",
-    "TBR", "1tHAvrhIaw_Gz47K5pL0SpeniEMYnczln", "1E7ucN5WOyynzNjeYTiPiS_ixbaJXA6cW",
-    "HR" , "1cByc4uVCf46lwVnvfECmKi3IOsR3T2sU", "1ca5tpyAAe2DzeiVStl4Cl8qEvwI7Jaxx",
-    "DGR", "1aSsCAHVwRyquDH_-tA89-7qObJ4eqc-v", "1aCvZAd7YqU5kzQfshmxIBNECWQqRJKoC",
-    "FR" , "1teWo-kp-ujtKtHfvoLfZZCKcGsh0twiD", "1lkGlWPqBDHHP8yN8U-lkxKBnkRKmfVgS") |>
+          "EDR", "1t_oT37auM_25SnU3TLrXa_xCyoV9H9qt", "1s-FhTTDlclwcuu-H-9wIwwC1K7lEijXI",
+          "TBR", "1tHAvrhIaw_Gz47K5pL0SpeniEMYnczln", "1E7ucN5WOyynzNjeYTiPiS_ixbaJXA6cW",
+          "HR" , "1cByc4uVCf46lwVnvfECmKi3IOsR3T2sU", "1ca5tpyAAe2DzeiVStl4Cl8qEvwI7Jaxx",
+          "DGR", "1aSsCAHVwRyquDH_-tA89-7qObJ4eqc-v", "1aCvZAd7YqU5kzQfshmxIBNECWQqRJKoC",
+          "FR" , "1teWo-kp-ujtKtHfvoLfZZCKcGsh0twiD", "1lkGlWPqBDHHP8yN8U-lkxKBnkRKmfVgS")
+
+yuba_mesh_data_by_reach <-
+  yuba_datasets |>
   mutate(filename_vertices = future_map_chr(vertices,
                                      function(x) drive_file_by_id(x, vsizip=T)),
          filename_thiessen = future_map_chr(thiessen,
@@ -123,21 +168,34 @@ yuba_mesh_prepped_by_reach <-
   mutate(geom_vertices = future_map(filename_vertices,
                              function(x) st_read(x, as_tibble=T)),
          geom_thiessen = future_map(filename_thiessen,
-                             function(x) st_read(x, as_tibble=T))) |>
-  mutate(mesh_prepped = future_pmap(list(geom_vertices, geom_thiessen, reach),
-                             function(vtx, thp, rch) {
-                               thp |>
-                                 st_join(vtx |>
-                                           transmute(vid=row_number()),
-                                         join=st_contains_properly, left = FALSE) |> #st_nearest_feature) |>
-                                 arrange(vid) |>
-                                 vector_prep_mesh(group_polygons = yuba_comid |> filter(reach==rch),
-                                                  .group_var = comid,
-                                                  .id_var = vid)
-                             })) |>
-  select(reach, mesh_prepped)
+                             function(x) st_read(x, as_tibble=T)))
 
-yuba_mesh_prepped_by_reach |> saveRDS(file.path(yuba_dir, "yuba_mesh_prepped_by_reach.Rds"))
+### ORIGINAL VERSION
+
+if(!file.exists(file.path(yuba_dir, "yuba_mesh_prepped_by_reach.Rds"))) {
+
+  yuba_mesh_prepped_by_reach <-
+    yuba_mesh_data_by_reach |>
+    mutate(mesh_prepped = future_pmap(list(geom_vertices, geom_thiessen, reach),
+                               function(vtx, thp, rch) {
+                                 thp |>
+                                   st_join(vtx |>
+                                             transmute(vid=row_number()),
+                                           join=st_contains_properly, left = FALSE) |> #st_nearest_feature) |>
+                                   arrange(vid) |>
+                                   vector_prep_mesh(group_polygons = yuba_comid |> filter(reach==rch),
+                                                    .group_var = comid,
+                                                    .id_var = vid)
+                               })) |>
+    select(reach, mesh_prepped)
+
+  yuba_mesh_prepped_by_reach |> saveRDS(file.path(yuba_dir, "yuba_mesh_prepped_by_reach.Rds"))
+
+} else {
+
+  yuba_mesh_prepped_by_reach <- readRDS(file.path(yuba_dir, "yuba_mesh_prepped_by_reach.Rds"))
+
+}
 
 yuba_output_by_reach <-
   inner_join(yuba_hsi_by_flow_by_reach, yuba_mesh_prepped_by_reach,
@@ -158,3 +216,89 @@ yuba_result <-
 outpath <- here::here("data-raw", "results", "fsa_yuba.Rds")
 
 yuba_result |> saveRDS(outpath)
+
+### VARIATION FOR REARING - BASEFLOW REMOVAL
+
+yuba_baseflow_poly <-
+  yuba_mesh_data_by_reach |>
+    mutate(result = future_pmap(list(geom_vertices, geom_thiessen, reach),
+                                function(vtx, thp, rch) {
+                                  thp |>
+                                    st_join(vtx |>
+                                              transmute(vid=row_number()),
+                                            join=st_contains_properly, left = FALSE) |>
+                                    arrange(vid) |>
+                                    filter(vid %in% filter(unnest(yuba_hsi_by_flow_by_reach, tbl_hsi), inundated & flow_cfs == 700 & reach == rch)$vid) |>
+                                    summarize() |>
+                                    st_union()
+                                })) |>
+  pull(result) |>
+  data.table::rbindlist() |>
+  st_sfc(crs = st_crs(yuba_comid)) |>
+  st_union()
+
+yuba_mesh_prepped_by_reach_nbfc <-
+  yuba_mesh_data_by_reach |>
+  mutate(mesh_prepped = future_pmap(list(geom_vertices, geom_thiessen, reach),
+                                    function(vtx, thp, rch) {
+                                      thp |>
+                                        st_join(vtx |>
+                                                  transmute(vid=row_number()),
+                                                join=st_contains_properly, left = FALSE) |> #st_nearest_feature) |>
+                                        arrange(vid) |>
+                                        vector_prep_mesh(group_polygons = yuba_comid |> filter(reach==rch),
+                                                         mask_polygons = yuba_baseflow_poly,
+                                                         .group_var = comid,
+                                                         .id_var = vid)
+                                    })) |>
+  select(reach, mesh_prepped)
+
+yuba_output_by_reach_nbfc <-
+  inner_join(yuba_hsi_by_flow_by_reach, yuba_mesh_prepped_by_reach_nbfc,
+             by=join_by(reach), relationship="one-to-one") |>
+  mutate(hsi_output = pmap(list(mesh_prepped, tbl_hsi),
+                           function(x, y) vector_summarize_hsi(x, y, .group_var=comid, .id_var=vid))) |>
+  mutate(hsi_final = pmap(list(hsi_output, reach),
+                          function(x, y) suitability_postprocess(x, yuba_comid |> filter(reach==y), .group_var=comid))) |>
+  select(reach, hsi_final)
+
+yuba_output_by_reach_nbfc |> saveRDS(file.path(yuba_dir, "yuba_output_by_reach_nbfc.Rds"))
+
+yuba_result_nbfc <-
+  yuba_output_by_reach_nbfc |>
+  select(reach, hsi_final) |>
+  unnest(hsi_final)
+
+outpath <- here::here("data-raw", "results", "fsa_yuba_nbfc.Rds")
+
+yuba_result_nbfc |> saveRDS(outpath)
+
+### VARIATION FOR SPAWNING - ALTERNATIVE HSI
+
+yuba_hsi_by_flow_by_reach_spawning <-
+  yuba_srh2d_result_by_reach |>
+  mutate(tbl_hsi = future_map(tbl_results,
+                              function(x) vector_calculate_hsi(x, hsi_func = vector_dvhsi_spawning))) |>
+  select(reach, tbl_hsi)
+
+yuba_hsi_by_flow_by_reach_spawning |> saveRDS(file.path(yuba_dir, "yuba_hsi_by_flow_by_reach_spawning.Rds"))
+
+yuba_output_by_reach_spawning <-
+  inner_join(yuba_hsi_by_flow_by_reach_spawning, yuba_mesh_prepped_by_reach,
+             by=join_by(reach), relationship="one-to-one") |>
+  mutate(hsi_output = pmap(list(mesh_prepped, tbl_hsi),
+                           function(x, y) vector_summarize_hsi(x, y, .group_var=comid, .id_var=vid))) |>
+  mutate(hsi_final = pmap(list(hsi_output, reach),
+                          function(x, y) suitability_postprocess(x, yuba_comid |> filter(reach==y), .group_var=comid))) |>
+  select(reach, hsi_final)
+
+yuba_output_by_reach_spawning |> saveRDS(file.path(yuba_dir, "yuba_output_by_reach_spawning.Rds"))
+
+yuba_result_spawning <-
+  yuba_output_by_reach_spawning |>
+  select(reach, hsi_final) |>
+  unnest(hsi_final)
+
+outpath <- here::here("data-raw", "results", "fsa_yuba_spawning.Rds")
+
+yuba_result_spawning |> saveRDS(outpath)
